requirements.txt
pydub>=0.25.1
numpy>=1.26.2
torch>=2.1.1
torchaudio>=2.2.1
requests>=2.31.0
ffmpeg-python>=0.2.0
openai-whisper
peewee>=3.17.1
pyannote.audio
beautifulsoup4>=4.12.3
bs4>=0.0.2
pydantic>=2.6.4
python-dotenv>=1.0.1
transformers>=4.39.2
accelerate>=0.29.3
whisperx>=3.1.5
spacy>=3.7.5